Added support for assigning float to stat
Created Stat#encodeFloat(float,int) to encode floats into fixed point int
Added maxhp stat to monster attributes
Changed debugging to output float stat

+++ core/src/com/riiablo/engine/server/Actioneer.java
@@ -123,7 +123,7 @@ public class Actioneer extends PassiveSystem {
         log.debug("{} attack {}", entityId, targetId);
 
         Attributes attrs = mAttributesWrapper.get(targetId).attrs;
-        log.debug("{} {}", targetId, attrs.get(Stat.hitpoints));
+        log.debug("{} {}", targetId, attrs.get(Stat.hitpoints).toFloat());
         break;
       case 27: // teleport
         mPosition.get(entityId).position.set(targetVec);
